+++ pysat/_instrument.py
@@ -229,12 +229,13 @@ class Instrument(object):
                 self.name = name.lower()
 
                 # Look to module for instrument functions and defaults
-                self._assign_attrs(by_name=True)
+                self._assign_attrs(by_name=True, tag=self.tag,
+                                   inst_id=self.inst_id)
             elif (platform is None) and (name is None):
                 # Creating "empty" Instrument object with this path
                 self.name = ''
                 self.platform = ''
-                self._assign_attrs()
+                self._assign_attrs(tag=self.tag, inst_id=self.inst_id)
             else:
                 raise ValueError(' '.join(('Inputs platform and name must both',
                                            'be strings, or both None.')))
@@ -251,7 +252,8 @@ class Instrument(object):
 
             # Look to supplied module for instrument functions and non-default
             # attribute values
-            self._assign_attrs(inst_module=self.inst_module)
+            self._assign_attrs(inst_module=self.inst_module,
+                               tag=self.tag, inst_id=self.inst_id)
 
         # More reasonable defaults for optional parameters
         self.clean_level = (clean_level.lower() if clean_level is not None
@@ -1095,7 +1097,8 @@ class Instrument(object):
         """
         pass
 
-    def _assign_attrs(self, by_name=False, inst_module=None):
+    def _assign_attrs(self, by_name=False, inst_module=None, tag=None,
+                      inst_id=None):
         """Assign all external instrument attributes to the Instrument object
 
         Parameters
@@ -1105,6 +1108,10 @@ class Instrument(object):
             if False uses inst_module. (default=False)
         inst_module : module or NoneType
             Instrument module or None, if not specified (default=None)
+        tag : str or NoneType
+            Instrument tag string
+        inst_id : str or NoneType
+            Instrument inst_id string
 
         Raises
         ------
@@ -1198,6 +1205,22 @@ class Instrument(object):
         else:
             # No module or name info, default pass functions assigned
             return
+
+        # Check if tag and inst_id are appropriate for the module
+        if inst_id not in inst.inst_ids.keys():
+            inst_id_str = ', '.join([ikey.__repr__()
+                                    for ikey in inst.inst_ids.keys()])
+            estr = ''.join(("'", inst_id, "' is not one of the supported ",
+                            'inst_ids. Supported inst_ids are: ',
+                            inst_id_str, '.'))
+            raise ValueError(estr)
+
+        if tag not in inst.inst_ids[inst_id]:
+            tag_str = ', '.join([tkey.__repr__()
+                                for tkey in inst.inst_ids[inst_id]])
+            estr = ''.join(("'", tag, "' is not one of the supported tags. ",
+                            'Supported tags are: ', tag_str, '.'))
+            raise ValueError(estr)
 
         # Assign the Instrument methods
         missing = list()